# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VR Park CRM API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Zone {
  id: number;
  name: string;
  color: string;
  capacity: number;
  /** @nullable */
  openTime?: string | null;
  /** @nullable */
  closeTime?: string | null;
  isActive?: boolean;
}

export interface ZoneInput {
  name: string;
  color: string;
  capacity: number;
  openTime?: string;
  closeTime?: string;
}

export interface ZoneUpdate {
  name?: string;
  color?: string;
  capacity?: number;
  openTime?: string;
  closeTime?: string;
  isActive?: boolean;
}

export interface SessionType {
  id: number;
  name: string;
  color: string;
  /** Minimum duration in minutes */
  minDuration: number;
}

export interface SessionTypeInput {
  name: string;
  color: string;
  minDuration: number;
}

export interface SessionTypeUpdate {
  name?: string;
  color?: string;
  minDuration?: number;
}

export interface Client {
  id: number;
  name: string;
  phone: string;
  /** @nullable */
  birthday?: string | null;
  /** @nullable */
  notes?: string | null;
  visitCount?: number;
  /** @nullable */
  lastVisit?: string | null;
}

export interface ClientInput {
  name: string;
  phone: string;
  birthday?: string;
  notes?: string;
}

export interface ClientUpdate {
  name?: string;
  phone?: string;
  birthday?: string;
  notes?: string;
}

export type BookingStatus = typeof BookingStatus[keyof typeof BookingStatus];


export const BookingStatus = {
  confirmed: 'confirmed',
  pending: 'pending',
  cancelled: 'cancelled',
  event: 'event',
} as const;

export interface Booking {
  id: number;
  /** @nullable */
  clientId?: number | null;
  /** @nullable */
  clientName?: string | null;
  /** @nullable */
  clientPhone?: string | null;
  /** @nullable */
  zoneId?: number | null;
  /** @nullable */
  zoneName?: string | null;
  /** @nullable */
  zoneColor?: string | null;
  /** @nullable */
  sessionTypeId?: number | null;
  /** @nullable */
  sessionTypeName?: string | null;
  /** @nullable */
  sessionTypeColor?: string | null;
  startTime: string;
  endTime: string;
  guestsCount: number;
  status: BookingStatus;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  adminName?: string | null;
}

export interface ClientDetail {
  id: number;
  name: string;
  phone: string;
  /** @nullable */
  birthday?: string | null;
  /** @nullable */
  notes?: string | null;
  visitCount?: number;
  /** @nullable */
  lastVisit?: string | null;
  bookings?: Booking[];
}

export type BookingInputStatus = typeof BookingInputStatus[keyof typeof BookingInputStatus];


export const BookingInputStatus = {
  confirmed: 'confirmed',
  pending: 'pending',
  cancelled: 'cancelled',
  event: 'event',
} as const;

export interface BookingInput {
  clientId?: number;
  clientName?: string;
  clientPhone?: string;
  zoneId?: number;
  sessionTypeId?: number;
  startTime: string;
  endTime: string;
  guestsCount: number;
  status: BookingInputStatus;
  notes?: string;
  adminName?: string;
}

export type BookingUpdateStatus = typeof BookingUpdateStatus[keyof typeof BookingUpdateStatus];


export const BookingUpdateStatus = {
  confirmed: 'confirmed',
  pending: 'pending',
  cancelled: 'cancelled',
  event: 'event',
} as const;

export interface BookingUpdate {
  clientId?: number;
  clientName?: string;
  clientPhone?: string;
  zoneId?: number;
  sessionTypeId?: number;
  startTime?: string;
  endTime?: string;
  guestsCount?: number;
  status?: BookingUpdateStatus;
  notes?: string;
  adminName?: string;
}

export type EventStatus = typeof EventStatus[keyof typeof EventStatus];


export const EventStatus = {
  planned: 'planned',
  active: 'active',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface Event {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  startTime: string;
  endTime: string;
  guestsCount: number;
  status: EventStatus;
  /** @nullable */
  zones?: string | null;
  /** @nullable */
  responsible?: string | null;
}

export type EventInputStatus = typeof EventInputStatus[keyof typeof EventInputStatus];


export const EventInputStatus = {
  planned: 'planned',
  active: 'active',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface EventInput {
  title: string;
  description?: string;
  startTime: string;
  endTime: string;
  guestsCount: number;
  status?: EventInputStatus;
  zones?: string;
  responsible?: string;
}

export type EventUpdateStatus = typeof EventUpdateStatus[keyof typeof EventUpdateStatus];


export const EventUpdateStatus = {
  planned: 'planned',
  active: 'active',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface EventUpdate {
  title?: string;
  description?: string;
  startTime?: string;
  endTime?: string;
  guestsCount?: number;
  status?: EventUpdateStatus;
  zones?: string;
  responsible?: string;
}

export interface EventStage {
  id: number;
  eventId: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  durationMinutes?: number | null;
  orderIndex: number;
}

export interface EventStageInput {
  name: string;
  description?: string;
  durationMinutes?: number;
  orderIndex: number;
}

export interface EventStageUpdate {
  name?: string;
  description?: string;
  durationMinutes?: number;
  orderIndex?: number;
}

export interface ZoneStat {
  zoneId: number;
  zoneName: string;
  zoneColor: string;
  bookingCount: number;
}

export interface DashboardSummary {
  totalBookings: number;
  confirmedBookings: number;
  pendingBookings: number;
  cancelledBookings: number;
  totalGuests: number;
  zoneStats: ZoneStat[];
  upcomingEvents?: number;
}

export type EventDetailStatus = typeof EventDetailStatus[keyof typeof EventDetailStatus];


export const EventDetailStatus = {
  planned: 'planned',
  active: 'active',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface EventDetail {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  startTime: string;
  endTime: string;
  guestsCount: number;
  status: EventDetailStatus;
  /** @nullable */
  zones?: string | null;
  /** @nullable */
  responsible?: string | null;
  stages?: EventStage[];
}

export type ListClientsParams = {
search?: string;
};

export type ListBookingsParams = {
/**
 * ISO date string (YYYY-MM-DD)
 */
date?: string;
zoneId?: number;
};

export type GetDashboardSummaryParams = {
date?: string;
};