# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VR Park CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List clients with optional search
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get client with booking history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client with bookings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientDetail"
        "404":
          description: Not found
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Bookings
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List bookings, filtered by date
      parameters:
        - name: date
          in: query
          schema:
            type: string
          description: ISO date string (YYYY-MM-DD)
        - name: zoneId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "409":
          description: Conflict with existing booking
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Not found
    patch:
      operationId: updateBooking
      tags: [bookings]
      summary: Update a booking (move, resize, status change)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "409":
          description: Conflict
    delete:
      operationId: deleteBooking
      tags: [bookings]
      summary: Delete a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Events
...
components:
  schemas:
...
    Client:
      type: object
      required: [id, name, phone]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        birthday:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        visitCount:
          type: integer
        lastVisit:
          type: ["string", "null"]
    ClientInput:
      type: object
      required: [name, phone]
      properties:
        name:
          type: string
        phone:
          type: string
        birthday:
          type: string
        notes:
          type: string
    ClientUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        birthday:
          type: string
        notes:
          type: string
    ClientDetail:
      type: object
      required: [id, name, phone]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        birthday:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        visitCount:
          type: integer
        lastVisit:
          type: ["string", "null"]
        bookings:
          type: array
          items:
            $ref: "#/components/schemas/Booking"
    Booking:
      type: object
      required: [id, startTime, endTime, guestsCount, status]
      properties:
        id:
          type: integer
        clientId:
          type: ["integer", "null"]
        clientName:
          type: ["string", "null"]
        clientPhone:
          type: ["string", "null"]
        zoneId:
          type: ["integer", "null"]
        zoneName:
          type: ["string", "null"]
        zoneColor:
          type: ["string", "null"]
        sessionTypeId:
          type: ["integer", "null"]
        sessionTypeName:
          type: ["string", "null"]
        sessionTypeColor:
          type: ["string", "null"]
        startTime:
          type: string
        endTime:
          type: string
        guestsCount:
          type: integer
        status:
          type: string
          enum: [confirmed, pending, cancelled, event]
        notes:
          type: ["string", "null"]
        adminName:
          type: ["string", "null"]
    BookingInput:
      type: object
      required: [startTime, endTime, guestsCount, status]
      properties:
        clientId:
          type: integer
        clientName:
          type: string
        clientPhone:
          type: string
        zoneId:
          type: integer
        sessionTypeId:
          type: integer
        startTime:
          type: string
        endTime:
          type: string
        guestsCount:
          type: integer
        status:
          type: string
          enum: [confirmed, pending, cancelled, event]
        notes:
          type: string
        adminName:
          type: string
    BookingUpdate:
      type: object
      properties:
        clientId:
          type: integer
        clientName:
          type: string
        clientPhone:
          type: string
        zoneId:
          type: integer
        sessionTypeId:
          type: integer
        startTime:
          type: string
        endTime:
          type: string
        guestsCount:
          type: integer
        status:
          type: string
          enum: [confirmed, pending, cancelled, event]
        notes:
          type: string
        adminName:
          type: string
...