# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VR Park CRM API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from './bookingStatus';

export interface Booking {
  id: number;
  /** @nullable */
  clientId?: number | null;
  /** @nullable */
  clientName?: string | null;
  /** @nullable */
  clientPhone?: string | null;
  /** @nullable */
  zoneId?: number | null;
  /** @nullable */
  zoneName?: string | null;
  /** @nullable */
  zoneColor?: string | null;
  /** @nullable */
  sessionTypeId?: number | null;
  /** @nullable */
  sessionTypeName?: string | null;
  /** @nullable */
  sessionTypeColor?: string | null;
  startTime: string;
  endTime: string;
  guestsCount: number;
  status: BookingStatus;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  adminName?: string | null;
}