# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VR Park CRM API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingInputStatus } from './bookingInputStatus';

export interface BookingInput {
  clientId?: number;
  clientName?: string;
  clientPhone?: string;
  zoneId?: number;
  sessionTypeId?: number;
  startTime: string;
  endTime: string;
  guestsCount: number;
  status: BookingInputStatus;
  notes?: string;
  adminName?: string;
}