# lib/api-zod/src/generated/types/clientInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VR Park CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface ClientInput {
  name: string;
  phone: string;
  birthday?: string;
  notes?: string;
}